Support impl blocks for types
This adds support for `impl Type` blocks to `maybe_async`,
`must_be_async` and `must_be_sync`.

The changes are minimal as this reuses the machinery already in place
for `impl Trait for Type`.

The main motivation for this change is better interoperability with
macros from other crates that apply to impl blocks and need
functions to be resolved to async/sync before they run.

Resolves #18